Provide UI for previously-added links
Re-adding a link will overwrite the previous one for the services we currently
support, so it's important to make that obvious to the user and avoid losing
previous tag data. We do that by prepopulating the tag input with any previous
tags (and checking the "private" checkbox when appropriate) and remind the user
of their previous bookmark in the feedback line.