Reduce OpenGL renderer-switching flicker
Somehow having a transition: all can cause a flicker when WebKit switches to
the OpenGL rendering engine. Replacing that with specific transition properties
seems to help. The interwebs suggested adding -webkit-backface-visibility:
hidden to the element, but that caused OpenGL rendering for the entire popup,
and since OpenGL doesn't handle antialiasing text it was ugly as sin.