Fix popup on the new-tab page
For some reason, the content script is allowed to run on the new tab
page (thus sending the request to enable the in-page popup), but Chrome
won't accept messages to that page, so the popup was not appearing at
all. Work around: just use the standard Chrome popup on chrome:// URLs.