Bugfix: clicking suggested tag
The code that was setting up a handler to restore the suggested tag to
the suggestion container upon removing it from the tag box was failing;
it had been depending on a value implicitly returned from a function to
which I had added another line of code. :/ Tests... someday? Sigh.